コメント記載。一部SpeakerのEnumからint64へ変更

+++ Sample/Plugins/VoicevoxEngine/Source/VoicevoxEngineEditor/Private/Components/SpeakerElementWidget.cpp
@@ -2,6 +2,7 @@
 
 
 #include "Components/SpeakerElementWidget.h"
+#include "VoicevoxCoreUtil.h"
 
 /**
  * @brief NativeConstruct override
@@ -10,7 +11,7 @@ void USpeakerElementWidget::NativeConstruct()
 {
 	if (SpeakerLabel)
 	{
-		SpeakerLabel->SetText(FText::AsCultureInvariant(FVoicevoxCoreUtil::GetMetaName(static_cast<int64>(Speaker))));
+		SpeakerLabel->SetText(FText::AsCultureInvariant(FVoicevoxCoreUtil::GetMetaName(Speaker)));
 	}
 
 }
